Extract apollo setup
