+++ resolvers/index.js
@@ -1,0 +1,21 @@
+const books = [
+  {
+    title: "Harry Potter and the Chamber of Secrets",
+    author: "J.K. Rowling",
+  },
+  {
+    title: "Jurassic Park",
+    author: "Michael Crichton",
+  },
+];
+
+// Resolvers define the technique for fetching the types defined in the
+// schema. This resolver retrieves books from the "books" array above.
+
+const resolvers = {
+  Query: {
+    books: () => books,
+  },
+};
+
+module.exports = resolvers;
